% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hellofunction}
\alias{hellofunction}
\title{Hello World Example}
\usage{
hellofunction(name)
}
\arguments{
\item{name}{a string that will appear before "Hello, world!"}
}
\description{
It's just an example to show how to develop a R package.
}
\examples{
\dontrun{
hello("Linda")
[1] "Linda Hello, world!"
}
}
\seealso{
\code{\link[base]{paste}}
}
